backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.0
python-multipart==0.0.6
websockets==12.0
pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6